requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
django-allauth==0.41.0
Django==3.0.4
django-extra-views==0.13.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-rest-passwordreset==1.1.0
factory-boy==2.12.0
Faker==4.0.2
honcho==1.0.1
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mock==4.0.2
oauthlib==3.1.0
pbr==5.4.4
phonenumbers==8.11.5
Pillow==7.0.0
psycopg2==2.8.4
purl==1.5
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
waitress==1.4.3
